# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Orbonomy",
    "version": "1.0.0",
    "description": "Unified AI-powered API platform with 50+ endpoints across accommodation, AI inference, content generation, fact-checking, and real-time data services.",
    "x-guidance": "POST to any endpoint with your data. The server returns 402 with payment-required info. Pay with USDC on Base chain and retry with the payment signature.",
    "contact": {
      "email": "orbonomy@proton.me"
    },
    "x-logo": {
      "url": "https://orbonomy.vercel.app/favicon.png"
    }
  },
  "servers": [
    {
      "url": "https://orbonomy.vercel.app"
    }
  ],
  "paths": {
    "/api/accommodation/search": {
      "post": {
        "operationId": "accommodation_search",
        "summary": "Search hotels, villas, and accommodations worldwide",
        "tags": [
          "Accommodation"
        ],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.050000"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "location": {
                    "type": "string",
                    "minLength": 1
                  },
                  "checkin": {
                    "type": "string",
                    "minLength": 1
                  },
                  "checkout": {
                    "type": "string",
                    "minLength": 1
                  },
                  "adults": {
                    "type": "integer"
                  },
                  "budget": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "location",
                  "checkin",
                  "checkout",
                  "adults",
                  "budget"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
...
    "/api/accommodation/compare": {
      "post": {
        "operationId": "accommodation_compare",
        "summary": "Compare prices across multiple booking sources",
        "tags": [
          "Accommodation"
        ],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.100000"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "location": {
                    "type": "string",
                    "minLength": 1
                  },
                  "checkin": {
                    "type": "string",
                    "minLength": 1
                  },
                  "checkout": {
                    "type": "string",
                    "minLength": 1
                  },
                  "budget": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "location",
                  "checkin",
                  "checkout",
                  "budget"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
    "/api/accommodation/nearby": {
      "post": {
        "operationId": "accommodation_nearby",
        "summary": "Find points of interest around a location",
        "tags": [
          "Accommodation"
        ],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.030000"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "location": {
                    "type": "string",
                    "minLength": 1
                  },
                  "radius": {
                    "type": "string",
                    "minLength": 1
                  },
                  "type": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "location",
                  "radius",
                  "type"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
    "/api/accommodation/reviews": {
      "post": {
        "operationId": "accommodation_reviews",
        "summary": "AI-summarized reviews for properties",
        "tags": [
          "Accommodation"
        ],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.050000"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "hotel_id": {
                    "type": "string",
                    "minLength": 1
                  },
                  "reviews_text": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "hotel_id",
                  "reviews_text"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
...
    "/api/accommodation/itinerary": {
      "post": {
        "operationId": "accommodation_itinerary",
        "summary": "Plan a full itinerary with accommodation and activities",
        "tags": [
          "Accommodation"
        ],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.250000"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "destination": {
                    "type": "string",
                    "minLength": 1
                  },
                  "days": {
                    "type": "integer"
                  },
                  "budget": {
                    "type": "string",
                    "minLength": 1
                  },
                  "interests": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "destination",
                  "days",
                  "budget",
                  "interests"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
...
    "/api/content/draft": {
      "post": {
        "operationId": "content_draft",
        "summary": "Generate first draft from brief or outline",
        "tags": [
          "Content Studio"
        ],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.150000"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string",
                    "minLength": 1
                  },
                  "tone": {
                    "type": "string",
                    "minLength": 1
                  },
                  "style": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "text",
                  "tone",
                  "style"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
...
    "/api/content/audit": {
      "post": {
        "operationId": "content_audit",
        "summary": "Content quality audit with scoring",
        "tags": [
          "Content Studio"
        ],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.080000"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "text"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
...
  }
}